move build loki entry to common area
