+++ internal/component/database_observability/loki_entry_test.go
@@ -1,4 +1,4 @@
-package collector
+package database_observability
 
 import (
 	"testing"
@@ -6,12 +6,11 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/grafana/alloy/internal/component/database_observability"
 	"github.com/grafana/alloy/internal/runtime/logging"
 )
 
 func TestBuildLokiEntry(t *testing.T) {
-	entry := buildLokiEntry(logging.LevelDebug, "test-operation", "test-instance", "This is a test log line")
+	entry := BuildLokiEntry(logging.LevelDebug, "test-operation", "test-instance", "This is a test log line")
 
 	require.Len(t, entry.Labels, 3)
 	require.Equal(t, database_observability.JobName, string(entry.Labels["job"]))
@@ -21,7 +20,7 @@ func TestBuildLokiEntry(t *testing.T) {
 }
 
 func TestBuildLokiEntryWithTimestamp(t *testing.T) {
-	entry := buildLokiEntryWithTimestamp(logging.LevelInfo, "test-operation", "test-instance", "This is a test log line", 5)
+	entry := BuildLokiEntryWithTimestamp(logging.LevelInfo, "test-operation", "test-instance", "This is a test log line", 5)
 
 	require.Equal(t, int64(5), entry.Entry.Timestamp.UnixNano())
 	require.Equal(t, time.Unix(0, 5), entry.Entry.Timestamp)
